+++ docs/src/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.3.1
 -------------------
 - Fixed critical error in vapor pressure to mixing ratio conversion function that caused incorrect results.
+- Fixed critical issue with documentation build
 
 1.3.0
 -------------------