+++ docs/src/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 - Fixed critical issue with documentation build
 - Added feature to `coated_wall_reactor` module to calculate fraction of unreacted surface sites after a given exposure time
 - Added support for accounting for flow around the outside of inserts in the `coated_wall_reactor` module when computing insert flow velocity
+- Change calculate_gamma to calculate_gamma_effective in `coated_wall_reactor` module.
 - Minor improvements
 
 1.3.0