+++ docs/src/changelog.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+1.3.1
+-------------------
+- Fixed critical error in vapor pressure to mixing ratio conversion function that caused incorrect results.
+- Fixed critical issue with documentation build
+- Added feature to `coated_wall_reactor` module to calculate fraction of unreacted surface sites after a given exposure time
+- Added support for accounting for flow around the outside of inserts in the `coated_wall_reactor` module when computing insert flow velocity
+- Change calculate_gamma to calculate_gamma_effective.
+- Minor improvements
+
 1.3.0
 -------------------
 - Improved documentation with usage examples and explanations of the API.